added logger usage
