+++ internal/api/responses.go
@@ -1,0 +1,29 @@
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+)
+
+func RespondWithJSON(w http.ResponseWriter, status int, payload interface{}) error {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(status)
+	return json.NewEncoder(w).Encode(payload)
+}
+func WriteError(ctx context.Context, w http.ResponseWriter, status int, message string) {
+	log := GetLogger(ctx)
+
+	err := RespondWithJSON(w, status, APIError{Message: message})
+	if err != nil {
+		log.WithError(err).Error("cannot write response")
+	}
+}
+func WriteSuccessful(ctx context.Context, w http.ResponseWriter, payload interface{}) {
+	log := GetLogger(ctx)
+
+	err := RespondWithJSON(w, http.StatusOK, payload)
+	if err != nil {
+		log.WithError(err).Error("cannot write response")
+	}
+}
